I2C maestro con libreria
