+++ .ccls-cache/@home@runner@Maestro-PIC-18F4550/main.c
@@ -1,14 +1,15 @@
 /*
- * File:   MainI2CIMU.c
+ * File:   MainI2C_Controlador.c
  * Author: mario
  *
- * Created on 20 de diciembre de 2021, 02:36 PM
+ * Created on 1 de Agosto de 2022, 09:24 AM
  */
 
 //Envió el cpodigo de I2C:
 
 #include<xc.h>
 #define _XTAL_FREQ 8000000
+#include "I2C.h"
 #pragma config FOSC=INTOSC_EC
 #pragma config WDT=OFF
 #pragma config PBADEN=OFF
@@ -16,15 +17,13 @@
 
 
 /*------------------------------ Funciones I2C ---------------------*/
-void Start(void);
-void Stop(void);
-void Rstart(void);
-unsigned char LecturaDato(unsigned char,unsigned char);
+
+
 /*------------------------------ Funciones para DS ---------------------*/
 unsigned char Recibir_Dato_DSPIC(unsigned char,unsigned char);
 void Enviar_Dato_DSPIC(unsigned char,unsigned char,unsigned char);
-void Enviar_Dato_I2C(unsigned char);
-unsigned char Recibir_Dato_I2C(unsigned char);
+
+
 
 /*------------------------------ variables del sistema -----------------*/
 char Vector_Datos[12] ;
@@ -41,12 +40,8 @@ void main(void){
   SSPCON2=0b00000000;
   SSPSTAT=0b11000000;
   SSPADD=19;
-  //PIR1=0;  
-  //__delay_ms(100);
   while(1){
     __delay_ms(10);
-      /*EnvioDato(0x0B,Contador,Contador);*/
-      //LATD=LecturaDatoA(0x10);
       if(1== Recibir_Dato_DSPIC(0x10,0x1)){
           LATC0=1;
       }
@@ -61,16 +56,14 @@ void main(void){
           LATC0=0;
       }
       LATD=Vector_Datos[0x0];
-      Enviar_Dato_DSPIC(0x10,0x11,90);
-      //LATD = LecturaDato(0x0A,2);
-    //EnvioDatoA(0x10,1);
-    //Contador++;
+      Enviar_Dato_DSPIC(0x10,0x11,50);
     LATC2=1;
     __delay_ms(1000);
     LATC2=0;
     __delay_ms(1000);
   }  
 }
+/*---------------------------- Funciones I2C --------------------------------*/
 void Enviar_Dato_DSPIC(unsigned char Direccion,unsigned char Registro, unsigned char Valor){
   char Verificador_Temp;
   /** Envio de Start **/
@@ -168,67 +161,6 @@ unsigned char Recibir_Dato_DSPIC(unsigned char Direccion , unsigned char Registr
 }
 
 
-void Enviar_Dato_I2C(unsigned char Dato){
-    while(RB2 == 1);
-    __delay_ms(250);
-    SSPBUF= Dato; //| 0b00000001; //& 0b11111110;
-    SSPIF=0;
-    while(SSPIF==0);
-}
-unsigned char Recibir_Dato_I2C(unsigned char Dato_Adicional){
-    while(RB2 == 1);
-    __delay_ms(500);
-    RCEN=1;
-    SSPIF=0;
-    while(SSPIF==0);
-    //LATD=SSPBUF;
-    //LATE2=1;
 
-    /** Envio de ACK **/
-    SSPIF=0;
-    if(Dato_Adicional==1) ACKDT=0;
-    else ACKDT=1;
-    ACKEN=1;
-    while(SSPIF==0);
-    //LATE2=0;
-    return SSPBUF;
-}
 
-void Start(void){    
-  SEN=1;
-  while(SEN==1);  
-}
-void Stop(void){
-  PEN=1;
-  while(PEN==1);  
-}
-void Rstart(void){
-  RSEN=1;
-  while(RSEN==1);  
-}
 
-unsigned char LecturaDato(unsigned char Direccion, unsigned char Registro){
-  Start();
-  SSPBUF=Direccion & 0b11111110;
-  SSPIF=0;
-  while(SSPIF==0);
-  SSPBUF=Registro;
-  SSPIF=0;
-  while(SSPIF==0);
-  Rstart();
-  SSPBUF=Direccion | 0b00000001;
-  SSPIF=0;
-  while(SSPIF==0);
-  RCEN=1;
-  SSPIF=0;
-  while(SSPIF==0);
-  
-  SSPIF = 0;
-  ACKDT = 1;
-  ACKEN = 1;
-  while(SSPIF==0);
-  
-  
-  Stop();
-  return SSPBUF;  
-}
